# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API HackR",
        "description": "Voici la documentation de l'API HackR. Vous pouvez tester les différentes routes (Authentification, hack, logs & rules) en utilisant l'interface Swagger.",
        "version": "1.0"
    },
    "paths": {
...
        "/api/changeRules": {
            "post": {
                "tags": [
                    "Droits"
                ],
                "summary": "Ajoute ou modifie les droits d'accès pour une fonctionnalité et un rôle",
                "description": "Cette méthode permet à un administrateur d'ajouter ou de modifier les droits d'accès pour une fonctionnalité donnée et un rôle spécifique. Si un droit existe déjà, il sera modifié, sinon il sera créé.",
                "operationId": "changeRules",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "fonctionnalite_id",
                                    "role_id"
                                ],
                                "properties": {
                                    "fonctionnalite_id": {
                                        "description": "L'ID de la fonctionnalité pour laquelle attribuer ou modifier les droits",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "role_id": {
                                        "description": "L'ID du rôle pour lequel attribuer ou modifier les droits",
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Droit ajouté ou modifié avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "string",
                                            "example": "Droit ajouté"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Droit déjà existant",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Droit déjà existant"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Utilisateur non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Utilisateur non authentifié"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès interdit pour l'utilisateur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Vous n'avez pas les droits pour accéder à cette ressource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur lors de la modification des droits",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Message d'erreur détaillé",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}